@@ -1,22 +1,86 @@
+//modules
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-
 import { AppRoutingModule } from './app-routing.module';
+import { HttpClientModule } from '@angular/common/http';
+import { FormsModule } from '@angular/forms';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { MatSliderModule } from '@angular/material/slider';
+import { MatIconModule } from '@angular/material/icon';
+import { MatButtonModule } from "@angular/material/button";
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatInputModule } from '@angular/material/input';
+import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatListModule } from '@angular/material/list';
+import { MatCardModule } from '@angular/material/card';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { OverlayModule } from '@angular/cdk/overlay';
+import { SlideshowModule } from 'ng-simple-slideshow';
+
+//components
 import { AppComponent } from './app.component';
-import { HomeComponent } from './components/home/home.component';
-import { HeaderComponent } from './components/header/header.component';
+import { HomeComponent } from './home/home.component';
+import { HeaderComponent } from './shared/components/header/header.component';
+import { BannerComponent } from './home/components/banner/banner.component';
+import { RankComponent } from './home/components/rank/rank.component';
+import { HotAlbumsComponent } from './home/components/hot-albums/hot-albums.component';
+import { FooterComponent } from './shared/components/footer/footer.component';
+import { SongCreateComponent } from './song/song-create/song-create.component';
+import { UserCreateComponent } from './user/user-create/user-create.component';
+import { PlayingComponent } from './playing/playing.component';
+import { SearchingComponent } from './searching/searching.component';
+import { LoginComponent } from './login/login.component';
+import { RegisterComponent } from './register/register.component';
+import { LoaderComponent } from './shared/components/loader/loader.component';
+
+//services
+import { SongService } from './shared/services/song.service';
+import { HotSingersComponent } from './home/components/hot-singers/hot-singers.component';
+import { SingerCreateComponent } from './singer/singer-create/singer-create.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     HomeComponent,
-    HeaderComponent
+    HeaderComponent,
+    BannerComponent,
+    RankComponent,
+    HotAlbumsComponent,
+    FooterComponent,
+    SongCreateComponent,
+    PlayingComponent,
+    SearchingComponent,
+    UserCreateComponent,
+    LoginComponent,
+    RegisterComponent,
+    LoaderComponent,
+    HotSingersComponent,
+    SingerCreateComponent
   ],
   imports: [
     BrowserModule,
-    AppRoutingModule
+    AppRoutingModule,
+    HttpClientModule,
+    FormsModule,
+    BrowserAnimationsModule,
+    MatSliderModule,
+    MatIconModule,
+    MatButtonModule,
+    MatToolbarModule,
+    MatInputModule,
+    MatSidenavModule,
+    MatListModule,
+    MatCardModule,
+    MatMenuModule,
+    MatProgressSpinnerModule,
+    OverlayModule,
+    SlideshowModule
   ],
-  providers: [],
+  entryComponents: [
+    LoaderComponent
+  ],
+  providers: [SongService],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
